+++ fynance/features/__init__.py
@@ -29,6 +29,7 @@ from . import (
     engineering,
     filters,
     garch,
+    horizon,
     indicators,
     momentums,
     money_management,
@@ -41,6 +42,7 @@ from . import (
 from .engineering import *
 from .filters import *
 from .garch import *
+from .horizon import *
 from .indicators import *
 from .momentums import *
 from .money_management import *
@@ -54,6 +56,7 @@ __all__ = engineering.__all__
 __all__ += regime.__all__
 __all__ += filters.__all__
 __all__ += garch.__all__
+__all__ += horizon.__all__
 __all__ += momentums.__all__
 __all__ += indicators.__all__
 __all__ += money_management.__all__